+++ yrs/src/tests/edit_traces_tests.rs
@@ -49,12 +49,12 @@ fn test_editing_trace(fpath: &str) {
             //let total = txt.len(&txn);
             if delete != 0 {
                 //println!("{at}/{total}: delete {delete_count} elements");
-                txt.remove_range(&mut txn, at as u32, delete as u32);
+                txt.remove_range(&mut txn, at as u64, delete as u64);
             }
             if !content.is_empty() {
                 //let len = content.len();
                 //println!("{at}/{total}: insert {len} elements - \"{content}\"");
-                txt.insert(&mut txn, at as u32, &content);
+                txt.insert(&mut txn, at as u64, &content);
             }
         }
     }
